# services/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Báscula API
  version: "1.0.0"
  description: |
    Multi-tenant farm labour and payroll API. One binary, one Postgres, four
    roles. Isolation is row level security keyed on `app.farm_id`, set from the
    access token inside the transaction that serves the request.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /v1/work-records:
    get:
      tags: [work-records]
      operationId: listWorkRecords
      summary: List work records
      description: |
        The weigher gets only the rows he recorded, and that narrowing is the
        RLS policy on `work_records`, not a WHERE clause in a handler.

        `q` matches the activity's name, the worker's name or the note.
      x-action: work_records.read
      x-roles: [owner, admin, weigher]
      parameters:
        - $ref: "#/components/parameters/QueryWorkerID"
        - name: activityId
          in: query
          schema: { type: string, format: uuid }
        - name: plotId
          in: query
          schema: { type: string, format: uuid }
        - name: plotCropId
          in: query
          schema: { type: string, format: uuid }
        - name: payScheme
          in: query
          schema: { $ref: "#/components/schemas/PayScheme" }
        - $ref: "#/components/parameters/QueryFrom"
        - $ref: "#/components/parameters/QueryTo"
        - $ref: "#/components/parameters/QueryQ"
        - $ref: "#/components/parameters/QueryStatus"
        - $ref: "#/components/parameters/QueryIncludeDeleted"
      responses:
        "200":
          description: Newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/WorkRecord" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [work-records]
      operationId: createWorkRecord
      summary: Record that somebody performed an activity
      description: |
        The one write in the system that decides money at write time, so most
        of it is about which price applies and when it froze:

          * `rateCents` given — the caller named the price, it freezes here,
            and a date range is legal;
          * the activity is priced by the week — the price is looked up when
            the settlement runs, which is exactly what the phone does today;
          * otherwise — the activity's rate in force on the day of the work,
            frozen now. NO_RATE_IN_FORCE when there is none.

        A record whose price is derived from a date must be a single day
        (decision 4); a range needs `rateCents`, and asking for both is 400
        with `details.code = RANGE_NEEDS_FROZEN_RATE`.

        A weigher may only record work priced by the week, and may not send a
        rate. He never sees a price and never needs to.
      x-action: work_records.write
      x-roles: [owner, admin, weigher]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkRecordInput" }
      responses:
        "200":
          description: That id already existed here.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkRecord" }
        "201":
          description: Recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkRecord" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: NO_RATE_IN_FORCE — that activity has no price on that date.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/work-records/{id}:
    parameters:
      - $ref: "#/components/parameters/PathID"
    get:
      tags: [work-records]
      operationId: getWorkRecord
      summary: One work record
      x-action: work_records.read
      x-roles: [owner, admin, weigher]
      responses:
        "200":
          description: The record.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkRecord" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [work-records]
      operationId: updateWorkRecord
      summary: Correct a work record that has not been paid
      description: |
        Exactly two things can change: the quantity and the note. The amount is
        recomputed here by the one money rule, `round(quantity × rate)` — a
        caller that could send its own total could send one that does not match
        its own line.

        What it refuses is the interesting half. The worker is somebody else's
        money. The date decides which rate period and which week apply, so
        moving it silently reprices the line. The activity decides the pay
        scheme, which a composite foreign key pins to the row. And the frozen
        price stays frozen: that is the answer to "why was I paid this". Any of
        those is a different work record — delete this one and write that one,
        in that order, so the anti double-pay lock sees both.
      x-action: work_records.admin
      x-roles: [owner, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkRecordPatch" }
      responses:
        "200":
          description: The record as it now stands.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkRecord" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            WORK_RECORD_SETTLED — a live settlement has claimed it. Work that
            has been paid is cancelled by voiding its settlement, never by
            editing it away underneath the payment.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [work-records]
      operationId: deleteWorkRecord
      summary: Remove a work record
      description: Logical, and refused on a settled record.
      x-action: work_records.admin
      x-roles: [owner, admin]
      responses:
        "204": { description: Marked deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: WORK_RECORD_SETTLED.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ---------------------------------------------------------------------------
  # The legacy pickup facade
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A 15-minute access token with claims `sub`, `farm_id`, `role`,
        `device_id`. The farm comes from here and from nowhere else.
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    QueryQ:
      name: q
      in: query
      description: Free-text needle, case-insensitive substring.
      schema: { type: string }
    QueryStatus:
      name: status
      in: query
      description: |
        `active` (the default), `inactive` for what has been logically removed,
        `all` for both.
      schema: { type: string, enum: [active, inactive, all] }
    QueryIncludeDeleted:
      name: includeDeleted
      in: query
      description: Legacy synonym for `status=all`, kept because the phone sends it.
      deprecated: true
      schema: { type: boolean }
...
    QueryFrom:
      name: from
      in: query
      schema: { type: string, format: date }
    QueryTo:
      name: to
      in: query
      schema: { type: string, format: date }
    QueryWorkerID:
      name: workerId
      in: query
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: BAD_REQUEST — malformed body, unknown field, or a bad value.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: UNAUTHORIZED — no token, or a token this server did not issue.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: FORBIDDEN — the permission table refused this role.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        NOT_FOUND. Also what a resource of another farm answers, and what an id
        that is not a UUID answers: nothing with that id exists here.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    ErrorCode:
      type: string
      description: |
        The machine-readable half of every failure. The client branches on
        this; the translation lives in the client, which is why `message` is
        English.

        This enum is compared against domain.AllCodes() by the contract test,
        so a code cannot exist in the server without appearing here.
      enum:
        - BAD_REQUEST
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - INTERNAL
        - TENANT_NOT_SET
        - INVALID_CREDENTIALS
        - EMAIL_NOT_VERIFIED
        - EMAIL_TAKEN
        - TOKEN_EXPIRED
        - TOKEN_REUSED
        - RATE_LIMITED
        - FARM_LIMIT_REACHED
        - FARM_SUSPENDED
        - WORK_RECORD_SETTLED
        - PAYABLE_ALREADY_CLAIMED
        - SETTLEMENT_ALREADY_VOID
        - ALREADY_REVERSED
        - NOTHING_TO_SETTLE
        - AMOUNT_EXCEEDS_BALANCE
        - INVALID_GEOMETRY
        - PLOT_HAS_ACTIVE_CROPS
        - NO_RATE_IN_FORCE
        - RANGE_NEEDS_FROZEN_RATE
        - DUPLICATE_DOCUMENT
        - DUPLICATE_NAME
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: English, for a developer and a log. Never shown to a picker.
            details:
              type: object
              additionalProperties: true
              description: |
                What a client needs to recover. The canonical case is
                PAYABLE_ALREADY_CLAIMED, where `winningSettlement` lets the
                phone re-derive its state.
...
    PayScheme:
      type: string
      description: |
        How an activity pays. Spanish values because they are the values stored
        in the Postgres enum, which the design documents fixed.
      enum: [contrato, tiempo, unidad_trabajo]
    RateSource:
      type: string
      description: |
        When the price froze. `explicit` — the caller named it, frozen at write
        time. `activity_dated` — the activity's rate in force on the record's
        day, frozen at write time. `weekly_price` — read from the week at
        settlement time, which is what the phone does today. The two derived
        forms require a single-day record.
      enum: [explicit, activity_dated, weekly_price]
...
    WorkRecord:
      type: object
      required: [id, workerId, activityId, payScheme, rateSource, quantity, dateFrom, dateTo]
      properties:
        id: { type: string, format: uuid }
        workerId: { type: string, format: uuid }
        activityId: { type: string, format: uuid }
        payScheme: { $ref: "#/components/schemas/PayScheme" }
        rateSource: { $ref: "#/components/schemas/RateSource" }
        startedAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time, nullable: true }
        dateFrom:
          type: string
          format: date-time
          description: The business day, in the farm's timezone, at midnight UTC.
        dateTo: { type: string, format: date-time }
        weekStart: { type: string, format: date-time, description: The Monday of dateFrom's ISO week. }
        quantity:
          type: string
          description: |
            A decimal with at most three places, sent as a JSON number and
            returned as a string. It is a string because a float64 round trip
            is exactly where two languages start disagreeing in the third
            decimal and nobody notices until payday.
        unitId: { type: string, format: uuid, nullable: true }
        rateCents:
          type: integer
          format: int64
          nullable: true
          description: Null while the price is the week's and has not been read yet.
        amountCents:
          type: integer
          format: int64
          nullable: true
          description: round(quantity × rateCents). Null for the same reason.
        note: { type: string, nullable: true }
        createdBy: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
        plotIds:
          type: array
          items: { type: string, format: uuid }
        plotCropIds:
          type: array
          items: { type: string, format: uuid }
        settled:
          type: boolean
          description: A live settlement has claimed it. It cannot be edited or removed.
    WorkRecordInput:
      type: object
      required: [activityId, workerId, quantity, dateFrom]
      properties:
        id: { type: string, format: uuid }
        activityId: { type: string, format: uuid }
        workerId: { type: string, format: uuid }
        quantity: { type: number, description: Positive; ignored and set to 1 for a contract. }
        rateCents:
          type: integer
          format: int64
          minimum: 1
          nullable: true
          description: |
            Freezes the price here and makes a date range legal. A weigher may
            not send one.
        dateFrom: { type: string, format: date }
        dateTo:
          type: string
          format: date
          description: |
            Defaults to `dateFrom`. A record with a date-derived price must be
            a single day.
        plotIds:
          type: array
          items: { type: string, format: uuid }
        plotCropIds:
          type: array
          items: { type: string, format: uuid }
        note: { type: string, nullable: true }
        deviceId: { type: string, nullable: true }
    WorkRecordPatch:
      type: object
      description: |
        The rejected fields are listed so the refusal has a sentence in it
        rather than a bare "malformed body": they are deliberate, not typos.
      properties:
        quantity: { type: number, description: Positive. The amount is recomputed. }
        note: { type: string, nullable: true }
        status: { type: string, enum: [active, inactive] }
        rateCents: { type: integer, format: int64, description: Rejected. }
        workerId: { type: string, format: uuid, description: Rejected. }
        activityId: { type: string, format: uuid, description: Rejected. }
        dateFrom: { type: string, format: date, description: Rejected. }
        dateTo: { type: string, format: date, description: Rejected. }
...